+++ framework/src/main/java/com/live2d/sdk/cubism/framework/motion/CubismMotionInternal.java
@@ -184,6 +184,7 @@ class CubismMotionInternal {
     /**
      * For strategy pattern.
      */
+    @FunctionalInterface
     public interface CsmMotionSegmentEvaluationFunction {
         float evaluate(final List<CubismMotionPoint> points, final float time);
     }